Move class RouteCollection into Foundation namespace
