+++ packages/framework/src/Hyde.php
@@ -5,6 +5,7 @@ namespace Hyde\Framework;
 use Hyde\Framework\Contracts\HydeKernelContract;
 use Hyde\Framework\Contracts\RouteContract;
 use Hyde\Framework\Foundation\PageCollection;
+use Hyde\Framework\Foundation\RouteCollection;
 use Hyde\Framework\Helpers\Features;
 use Illuminate\Support\Facades\Facade;
