+++ packages/framework/tests/Feature/RouteCollectionTest.php
@@ -8,12 +8,12 @@ use Hyde\Framework\Models\Pages\DocumentationPage;
 use Hyde\Framework\Models\Pages\MarkdownPage;
 use Hyde\Framework\Models\Pages\MarkdownPost;
 use Hyde\Framework\Models\Route;
-use Hyde\Framework\RouteCollection;
+use Hyde\Framework\Foundation\RouteCollection;
 use Hyde\Testing\TestCase;
 use Illuminate\Support\Collection;
 
 /**
- * @covers \Hyde\Framework\RouteCollection
+ * @covers \Hyde\Framework\Foundation\RouteCollection
  */
 class RouteCollectionTest extends TestCase
 {
